// src/lib/components/NavigationSidebar/types/selected-node.ts
/*
 * AtroCore Software
 *
 * This source file is available under GNU General Public License version 3 (GPLv3).
 * Full copyright and license information is available in LICENSE.txt, located in the root directory.
 *
 * @copyright  Copyright (c) AtroCore GmbH (https://www.atrocore.com)
 * @license    GPLv3 (https://www.gnu.org/licenses/)
 */

export type SelectedNode = {
    id: string;
    name: string;
    scope: string;
    link: string;
    icon?: string | null;
}